Rename `arena.rs` to `index_vec.rs`
