+++ crates/oak_index/src/lib.rs
@@ -1,3 +1,3 @@
-pub(crate) mod arena;
 pub mod builder;
+pub(crate) mod index_vec;
 pub mod semantic_index;
